+++ cppbuild/src/Cluster.hpp
@@ -9,29 +9,29 @@ class Cluster : public LocalizedObject
 
 protected:
   std::vector<LocalizedObject::Point> m_points;
-  uint16_t m_peak;
-  uint32_t m_total;
+  int m_peak;
+  int m_total;
 
 public:
   Cluster();
   virtual ~Cluster();
   Cluster* getCopy();
-  void addPoint(uint16_t x, uint16_t y);
+  void addPoint(int x, int y);
   void addPoint(LocalizedObject::Point pt){ m_points.push_back(pt); }
-  void addPoint(uint16_t x, uint16_t y, uint16_t i);
-  void addPoint(LocalizedObject::Point pt, uint16_t i);
-  LocalizedObject::Point getPoint(uint32_t index){ return m_points.at(index); }
-  void setPoint(uint32_t index, LocalizedObject::Point pt){ m_points.at(index) = pt; }
-  void removePoint(uint32_t index){ m_points.erase(m_points.begin()+index); }
+  void addPoint(int x, int y, int i);
+  void addPoint(LocalizedObject::Point pt, int i);
+  LocalizedObject::Point getPoint(int index){ return m_points.at(index); }
+  void setPoint(int index, LocalizedObject::Point pt){ m_points.at(index) = pt; }
+  void removePoint(int index){ m_points.erase(m_points.begin()+index); }
   void removePoint(LocalizedObject::Point pt);
   void clearPoints(){ m_points.clear(); }
-  void setPeakIntensity(uint16_t i){ m_peak = i; }
-  uint16_t peak(){ return m_peak; }
-  void setIntegratedIntensity(uint32_t i){ m_total = i; }
-  uint32_t integratedIntensity(){ return m_total; }
-  uint32_t indexOf(LocalizedObject::Point pt);
-  uint32_t size(){ return m_points.size(); }
-  uint32_t getBorderLength(Cluster* c);
+  void setPeakIntensity(int i){ m_peak = i; }
+  int peak(){ return m_peak; }
+  void setIntegratedIntensity(int i){ m_total = i; }
+  int integratedIntensity(){ return m_total; }
+  int indexOf(LocalizedObject::Point pt);
+  int size(){ return m_points.size(); }
+  int getBorderLength(Cluster* c);
   void add(Cluster* c);
   double peakToPeakDistance2(Cluster* c);
   Mask* getMask(int width, int height, bool outline=false);
