+++ cppbuild/src/FileManager.cpp
@@ -11,18 +11,18 @@ FileManager::~FileManager()
 
 ImSeries* FileManager::load()
 {
-  uint8_t np = 0;
-  uint8_t nt = 0;
+  int np = 0;
+  int nt = 0;
   std::vector<ImStack*> stacks;
-  uint8_t nx[4] = {1,1,1,1};
+  int nx[4] = {1,1,1,1};
   for(m_it = m_fileList.begin(); m_it != m_fileList.end(); m_it++){
-    for(uint8_t i = 0; i < m_it->np*m_it->nt; i++) stacks.push_back(new ImStack(m_it->nw,m_it->nz));
+    for(int i = 0; i < m_it->np*m_it->nt; i++) stacks.push_back(new ImStack(m_it->nw,m_it->nz));
     Dimension* order = m_it->order;
     nx[WAVELENGTH] = m_it->nw;
     nx[ZSLICE] = m_it->nz;
     nx[POSITION] = np + m_it->np;
     nx[TIMEPOINT] = nt + m_it->nt;
-    uint8_t ix[4] = {0,0,np,nt};
+    int ix[4] = {0,0,np,nt};
     np += m_it->np;
     nt += m_it->nt;
     for(std::vector<std::string>::iterator file_it = m_it->fnames.begin(); file_it != m_it->fnames.end(); file_it++){
@@ -61,14 +61,14 @@ ImSeries* FileManager::loadNext()
   if(m_it == m_fileList.end()) return NULL;
   nia::nout << "Loading image series " << m_it->sname << "\n";
   std::vector<ImStack*> stacks;
-  for(uint8_t i = 0; i < m_it->np*m_it->nt; i++) stacks.push_back(new ImStack(m_it->nw,m_it->nz));
+  for(int i = 0; i < m_it->np*m_it->nt; i++) stacks.push_back(new ImStack(m_it->nw,m_it->nz));
   Dimension* order = m_it->order;
-  uint8_t nx[4] = {1,1,1,1};
+  int nx[4] = {1,1,1,1};
   nx[WAVELENGTH] = m_it->nw;
   nx[ZSLICE] = m_it->nz;
   nx[POSITION] = m_it->np;
   nx[TIMEPOINT] = m_it->nt;
-  uint8_t ix[4] = {0,0,0,0};
+  int ix[4] = {0,0,0,0};
   for(std::vector<std::string>::iterator file_it = m_it->fnames.begin(); file_it != m_it->fnames.end(); file_it++){
     std::vector<ImFrame*> frames = ImFrame::load((*file_it).c_str());
     for(std::vector<ImFrame*>::iterator fit = frames.begin(); fit != frames.end(); fit++){
@@ -102,18 +102,18 @@ ImSeries* FileManager::loadNext()
 
 ImSeries* FileManager::loadMT()
 {
-  uint8_t np = 0;
-  uint8_t nt = 0;
+  int np = 0;
+  int nt = 0;
   std::vector<ImStack*> stacks;
-  uint8_t nx[4] = {1,1,1,1};
+  int nx[4] = {1,1,1,1};
   for(m_it = m_fileList.begin(); m_it != m_fileList.end(); m_it++){
-    for(uint8_t i = 0; i < m_it->np*m_it->nt; i++) stacks.push_back(new ImStack(m_it->nw,m_it->nz));
+    for(int i = 0; i < m_it->np*m_it->nt; i++) stacks.push_back(new ImStack(m_it->nw,m_it->nz));
     Dimension* order = m_it->order;
     nx[WAVELENGTH] = m_it->nw;
     nx[ZSLICE] = m_it->nz;
     nx[POSITION] = np + m_it->np;
     nx[TIMEPOINT] = nt + m_it->nt;
-    uint8_t ix[4] = {0,0,np,nt};
+    int ix[4] = {0,0,np,nt};
     np += m_it->np;
     nt += m_it->nt;
     std::vector< std::vector<ImFrame*>* > frames;
